copy and paste blade code into new vue component
